+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 This project follows [Semantic Versioning](http://semver.org/).
 
+## [36.2.0] 2024-07-03
+### Changed
+- addFormToPage() can optionally take a questionnaire response in options
+  to reduce number of lforms API calls.
+- Error message for missing enableWhen.question.
+
 ## [36.1.3] 2024-06-09
 ### Fixed
 - Removed accidental creation of global variable "self", which caused Google